Add API endpoints, change URLS, make ID visible
* Add endpoints for confirm_ships and get_state. Note that the endpoints
are defined differently than our initial ideas in the meeting notes.

* Change URLs to have dashes instead of underscores for endpoints, in
order to follow typical URL style guidelines. The endpoint names and
Python functions are unchanged; in keeping with Python style guidelines,
they still have underscores.

* Make ID field of Game/Player/Board objects visible for admin (read-only).

* Define fire_shot endpoint, which will be completed later today.

Endpoint examples:

* name: new_game
* path: new-game/<int:player1_id>/<int:player2_id>/<int:num_ships>
* example: http://localhost:8000/play/new-game/3/4/7
* response: {"game_id": 12}

The example above creates a new game between a player with an ID of 3
and a player with an ID of 4. The game will have 7 ships. We can see
that the game ID is 12.

* name: confirm_ships
* path: confirm-ships/<int:game_id>/<int:player_id>/<ship_board>
* example: http://localhost:8000/play/confirm-ships/12/4/-----------a---
------a------------bbbb----------------c---------c---------c------------
--------ddddd
* response: {"ship_board": "-----------a---------a------------bbbb------
----------c---------c---------c--------------------ddddd"}

The example above confirms the ship placement for the player with an ID
of 4 in the game with an ID of 12. The string representing the board
has been saved and returned.

* name: get_state
* path: get-state/<int:game_id>/<int:player_id>/<is_my_board>
* example: http://localhost:8000/play/get-state/12/4/true
* response: {"attack_board": "", "ship_board": "-----------a---------a--
----------bbbb----------------c---------c---------c--------------------d
dddd"}

The example above returns the board states for the player with an ID of
4 in the game with an ID of 12. Since is_my_board is true, it returns
the player's own attack board and ship board.